move sha384 to another file
