+++ CMakeLists.txt
@@ -93,7 +93,7 @@ file(GLOB_RECURSE LIB_SRC
         #########
         ${CMAKE_CURRENT_SOURCE_DIR}/app/src/b64url.c
         ${CMAKE_CURRENT_SOURCE_DIR}/app/src/parser.c
-        ${CMAKE_CURRENT_SOURCE_DIR}/app/src/crypto.c
+        ${CMAKE_CURRENT_SOURCE_DIR}/app/src/crypto_helper.c
         ${CMAKE_CURRENT_SOURCE_DIR}/app/src/parser_impl.c
         )
