+++ app/src/crypto_helper.c
@@ -1,5 +1,5 @@
 /*******************************************************************************
-*   (c) 2019 Zondax GmbH
+*   (c) 2018 - 2022 Zondax AG
 *
 *  Licensed under the Apache License, Version 2.0 (the "License");
 *  you may not use this file except in compliance with the License.
@@ -13,27 +13,25 @@
 *  See the License for the specific language governing permissions and
 *  limitations under the License.
 ********************************************************************************/
-
-#pragma once
-
-#ifdef __cplusplus
-extern "C" {
+#if defined(LEDGER_SPECIFIC)
+    #include "bolos_target.h"
+    #include "cx.h"
+#else
+    #include <sha512.h>
 #endif
 
-#include <zxmacros.h>
-#include "coin.h"
-#include <stdbool.h>
-#include <sigutils.h>
-#include "zxerror.h"
+#include "zxmacros.h"
 
-zxerr_t crypto_fillAddress(uint8_t *buffer, uint16_t bufferLen, uint16_t *addrLen);
-
-zxerr_t crypto_getsignature_part(uint8_t *buffer, uint16_t bufferLen, uint8_t index);
-zxerr_t crypto_getpubkey_part(uint8_t *buffer, uint16_t bufferLen, uint8_t index);
-
-zxerr_t crypto_sign(uint8_t *signature, uint16_t signatureMaxlen,
-                    const uint8_t *message, uint16_t messageLen, uint16_t *sigSize);
-
-#ifdef __cplusplus
+void crypto_sha384(const unsigned char *in, unsigned int inLen, unsigned char *out, unsigned int outLen) {
+#if defined(TARGET_NANOS) || defined(TARGET_NANOX) || defined(TARGET_NANOS2)
+    cx_sha512_t ctx;
+    cx_sha384_init(&ctx);
+    cx_hash(&ctx.header, CX_LAST, in, inLen, out, outLen);
+#else
+    uint8_t tmp[64] = {0};
+    // This function requires a full 64 bytes context (sha512)
+    SHA384(in, inLen, tmp);
+    // We can later trim and return only 48
+    MEMCPY(out, tmp, 48);
+#endif
 }
-#endif
